ONNX/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.9
insightface>=0.7.3
onnxruntime>=1.16.0